requirements.txt
Django~=4.1.0
Pillow==9.2.0
django-braces==1.15.0
django-embed-video==1.4.4
pymemcache==3.5.2
django-debug-toolbar==3.6.0
redis==4.3.4
django-redisboard==8.3.0
attrs
djangorestframework==3.13.1